MDL update, improved fullscreen
Upgraded mdl to 1.0.6, made a fullscreen toggle button appear on the map
when in fullscreen mode. When launched from "add to home screen" in
android the fullscreen back button exits out of the app entirely, so
needed something better to click there. Also fixed react-dom deprecation
warning so it should be ready for 0.15.